Remove Encounter move fetchNotes
