+++ src/data/fetchNotes.js
@@ -1,15 +1,16 @@
-/**
- * This function will 
- * 1. Fetch all the notes from the database
- * 2. Fetch the text content of each note and create a ClinicalNote object for each note
- * 3. Return an array of ClinicalNote objects
- * 4. The function will return an array of ClinicalNote objects and the number of notes generated
- */
-
 import ClinicalNote from '@/models/ClinicalNote';
 
-async function fetchNotes(client, patientId) {
+export default async function fetchNotes(client, patientId) {
+    /**
+     * This function will 
+     * 1. Fetch all the notes from the database
+     * 2. Fetch the text content of each note and create a ClinicalNote object for each note
+     * 3. Return an array of ClinicalNote objects
+     * 4. The function will return an array of ClinicalNote objects and the number of notes generated
+     */
+
     let noteSearchData;
+    
     try {
         noteSearchData = await client.request("/DocumentReference?patient=" + patientId);
     } catch (error) {
@@ -65,4 +66,3 @@ function pullTextContent(html) {
     textClean = textClean.replace(/[^a-zA-Z0-9\s]/g, '');
     return textClean;
 }
-export default fetchNotes;
